requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.7.0
asyncpg==0.30.0
bcrypt==4.2.1
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
h11==0.14.0
idna==3.10
iso8601==2.1.0
pydantic==2.10.3
pydantic-settings==2.7.0
pydantic_core==2.27.1
pypika-tortoise==0.3.2
python-dotenv==1.0.1
pytz==2024.2
sniffio==1.3.1
starlette==0.41.3
tortoise-orm==0.22.2
typing_extensions==4.12.2
uvicorn==0.32.1
